requirements.txt
django-celery>=3.0
